test(check): split import scanner regressions
